requirements.txt
Django==5.1.3
django-ninja==1.1.0